Add integration tests (#9)
Integration tests based on docker compose and a test shell script. A
custom multi-stage docker image is used to build the lastest version of
gh-get. Then an image that has GitHub CLI preinstalled is used for
running the tests. The test script first installs the prebuild gh-get
binary. For this to work, GitHub CLI needs to be authenticated. This
requires passing the GH_TOKEN environment variable into the test docker
container. The tests cover some of the happy path scenarios.